+++ src/main/java/org/programers/vouchermanagement/member/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package org.programers.vouchermanagement.member.dto;
+package org.programers.vouchermanagement.member.dto.response;
 
 import org.programers.vouchermanagement.member.domain.Member;
