+++ src/main/java/org/programers/vouchermanagement/member/dto/response/MembersResponse.java
@@ -1,4 +1,4 @@
-package org.programers.vouchermanagement.member.dto;
+package org.programers.vouchermanagement.member.dto.response;
 
 import java.util.List;
