+++ src/main/java/org/programers/vouchermanagement/voucher/dto/response/VouchersResponse.java
@@ -1,4 +1,4 @@
-package org.programers.vouchermanagement.voucher.dto;
+package org.programers.vouchermanagement.voucher.dto.response;
 
 import java.util.List;
