+++ src/main/java/org/programers/vouchermanagement/wallet/dto/request/WalletCreationRequest.java
@@ -1,4 +1,4 @@
-package org.programers.vouchermanagement.wallet.dto;
+package org.programers.vouchermanagement.wallet.dto.request;
 
 import java.util.UUID;
