+++ src/main/java/org/programers/vouchermanagement/wallet/dto/response/WalletsResponse.java
@@ -1,4 +1,4 @@
-package org.programers.vouchermanagement.wallet.dto;
+package org.programers.vouchermanagement.wallet.dto.response;
 
 import java.util.List;
